refactor: move token_verification_metadata endpoints to dedicated api file
