+++ propelauth_py/__init__.py
@@ -31,9 +31,11 @@ from propelauth_py.api_v2.org import (
     _disallow_org_to_setup_saml_connection,
 )
 from propelauth_py.api_v2.magic_link import _create_magic_link
-from propelauth_py.api import (
+from propelauth_py.api_v2.token_verification_metadata import (
     _fetch_token_verification_metadata,
     TokenVerificationMetadata,
+)
+from propelauth_py.api import (
     OrgQueryOrderBy,
     UserQueryOrderBy,
     _migrate_user_from_external_source,
